v2.4.18-pre4 -> v2.4.18-pre5
  - Include missing radeonfb defines
  - Fix fs/buffer.c thinko introduced in pre4
  - USB bugfixes
  - Make fat work correctly with gcc-3.0.x
  - Avoid overusage of the vmalloc area by
    NTFS
  - atyfb: Decrease clock rate for 3d RAGE XL
  - Sungem driver bugfixes
  - More networking updates
  - More SPARC updates
  - devfs update
  - Reiserfs expanding truncate fix
  - ext3 update
  - Add support to WDIOC_SETTIMEOUT on several
    watchdog drivers
  - dl2k driver update
  - Orinoco driver update
  - Radeonfb driver update
  - Avoid free_swap_and_cache() from leaving
    freeable pages on the cache
  - Add workarounds for AMD Elan processors
  - Random pmac driver bugfixing
  - emu10k1 driver update

+++ fs/intermezzo/presto.c
@@ -190,7 +190,7 @@ int presto_chk(struct dentry *dentry, int flag)
         }
 
         EXIT;
-        return ((int)dentry->d_fsdata & flag);
+        return ((int)(long)dentry->d_fsdata & flag);
 }
 
 /* set a bit in the dentry flags */
@@ -211,7 +211,7 @@ void presto_set(struct dentry *dentry, int flag)
                 }
         } else {
                 CDEBUG(D_INODE, "Setting dentry->d_fsdata\n");
-                ((int)dentry->d_fsdata) |= flag;
+                ((int)(long)dentry->d_fsdata) |= flag;
         }
 }
 
@@ -491,10 +491,10 @@ int presto_mark_dentry(const char *name, int and_flag, int or_flag,
                 CDEBUG(D_INODE, "fset %p, flags %x data %x\n", 
                        fset, fset->fset_flags, fset->fset_data);
         } else {
-                ((int)dentry->d_fsdata) &= and_flag;
-                ((int)dentry->d_fsdata) |= or_flag;
+                ((int)(long)dentry->d_fsdata) &= and_flag;
+                ((int)(long)dentry->d_fsdata) |= or_flag;
                 if (res) 
-                        *res = (int)dentry->d_fsdata;
+                        *res = (int)(long)dentry->d_fsdata;
         }
 
         /* indicate if we were the only users while changing the flag */
